% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redisconnect.R
\name{connect_redis}
\alias{connect_redis}
\title{Connect to redis database for direct write access}
\usage{
connect_redis(host = ifelse(.ahubEnv$debug, .ahubEnv$debughost, "redis"))
}
\arguments{
\item{host}{}
}
\description{
Connect to redis database for direct write access
}
