% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{dummy_process}
\alias{dummy_process}
\title{Dummy process for demonstrating AHUB batch and thread functionality}
\usage{
dummy_process(t = 10, steps = 5)
}
\arguments{
\item{t}{(float) total process time}

\item{steps}{(integer) How many process steps}
}
\value{
NOthing
}
\description{
Dummy process for demonstrating AHUB batch and thread functionality
}
