% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bosswrapper.R
\name{get_current_pid}
\alias{get_current_pid}
\title{Retrieve pid from given environment}
\usage{
get_current_pid(envir)
}
\arguments{
\item{envir}{environment}
}
\description{
Retrieve pid from given environment
}
