% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importswagger.R
\name{get_node_api}
\alias{get_node_api}
\title{Get Node API from swagger documentation}
\usage{
get_node_api(dir = "node1", host = "localhost", port = "8000",
  uid = NULL, pwd = NULL)
}
\arguments{
\item{pwd}{}
}
\description{
Get Node API from swagger documentation
}
