% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bossconnect.R
\name{init_boss_api}
\alias{init_boss_api}
\title{Get the boss api object}
\usage{
init_boss_api(host = ifelse(.ahubEnv$debug, .ahubEnv$debughost, "nginx"),
  port = 8000)
}
\arguments{
\item{host}{}
}
\description{
Get the boss api object
}
