% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchprocess.R
\name{init_future}
\alias{init_future}
\title{Initiate multisession future for batch process}
\usage{
init_future()
}
\value{
TRUE/FALSE if successful
}
\description{
Initiate multisession future for batch process
}
