% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bosswrapper.R
\name{pid_log}
\alias{pid_log}
\title{Write log to boss with given PID}
\usage{
pid_log(msg, level = "INFO")
}
\arguments{
\item{msg}{Message to log}

\item{level}{Log level INFO, DEBUG, WARN, or ERROR}
}
\value{
Nothing
}
\description{
Write log to boss with given PID
}
