% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadprocess.R
\name{thread_process}
\alias{thread_process}
\title{Title}
\usage{
thread_process(myfun, process_name = "thread", arglist, debug = F)
}
\arguments{
\item{myfun}{function, which is to be executed in thread mode}

\item{process_name}{name of the process to be saved}

\item{arglist}{named list of arguments for myfun}

\item{debug}{TRUE/FALSE endpoint for call to boss is switched to .ahubEnv$debughost instead of "nginx"}
}
\value{
list with message, logitems, status and process id
}
\description{
Title
}
\examples{
\dontrun{
thread_process(dummy_process, "test", arglist = list(t=1))
}
}
