% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchprocess.R
\name{daily_batch_process}
\alias{daily_batch_process}
\title{Title}
\usage{
daily_batch_process(myfun, process_name = "batch", arglist = list(),
  force = F, debug = F)
}
\arguments{
\item{arglist}{}
}
\description{
Title
}
