% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bossconnect.R
\name{init_boss_api}
\alias{init_boss_api}
\title{Get the boss api object}
\usage{
init_boss_api(host = ifelse(.ahubEnv$debug, .ahubEnv$debughost,
  .ahubEnv$host), port = 8000)
}
\arguments{
\item{host}{(character) hostname}

\item{port}{(int) port}
}
\value{
TRUE/FALSE if connection was successfully established
}
\description{
Get the boss api object
}
